+++ code/ruby/shared/air.rb
@@ -1,0 +1,2 @@
+module Air
+end
